+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+0.0.5 - 2022-04-20
+==================
+### Refactoring
+ - use kptncook api client instead of repository like mealie
+    - #11 issue by @ephes / @gloriousDan
+### Features
+ - new cli command `kptncook kptncook-access_token` fetches the access token from the kptncook api
+    - #11 issue by @ephes / @gloriousDan
+ - new cli command `kptncook list-recipes` lists all locally stored recipes
+    - #11 issue by @ephes / @gloriousDan
+ - new cli command `kptncook backup-favorites` fetches all favorites from the kptncook api and stores them locally
+    - #11 issue by @ephes / @gloriousDan
+ - new cli command `kptncook search-by-id` searches for a recipe by id (url, uid or oid) and stores it locally
+    - #11 issue by @ephes / @gloriousDan
+
 0.0.4 - 2022-04-08
 ==================
 ### Fixes